Decouple Linux VT
Move Linux VT into separate files to enable developers to use it in
other backends, such as Linux DRM.
